pyproject.toml
[build-system]
requires = ["hatchling>=1.25", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[project]
name = "llm-telemetry-toolkit"
dynamic = ["version"]
description = "A standardized, file-based telemetry toolkit for LLM Agents."
readme = "README.md"
authors = [
    {name = "Roy Dawson IV", email = "roy.dawson.iv@gmail.com"},
]
license = "MIT"
requires-python = ">=3.8"
keywords = ["llm", "logging", "telemetry", "observability", "agent", "pydantic", "json", "markdown"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "rich>=10.0",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-http>=1.24",
]

[project.scripts]
llm-telemetry-toolkit = "llm_telemetry_toolkit.interface.cli:main"

[project.urls]
Homepage = "https://github.com/imyourboyroy/llm-telemetry-toolkit"
Repository = "https://github.com/imyourboyroy/llm-telemetry-toolkit"
"Bug Tracker" = "https://github.com/imyourboyroy/llm-telemetry-toolkit/issues"

[tool.hatch.version]
source = "vcs"
tag-pattern = "v(?P<version>.*)"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_telemetry_toolkit"]
